Weather Page
These are the HTML, CSS and JS files from my weather page. So far, the searchbar is working and displaying the maximum and minimum temperature of the day. Also, the gps button is working, if the user agrees to share his position.

Still missing: the weather forecast. And I still need to update the time to be the time from the searched place or location.